Support mapping docker scopes to container names
Adds flag `--metrics.docker-scopes` which, when enabled, attempts to map
any cgroup paths with name "docker-{id}.scope" to the corresponding
container name, using `docker inspect`.

Scopes are likely hard-coded to be filtered due to their ephemeral
nature, and in this case, the high cardinality of changing container
IDs. This is addressed by mapping it to an actual image name, which is
at least more useful than simply dropping all scope cgroups.